error page and move util function
