+++ src/utils.ts
@@ -132,4 +132,17 @@ export enum PostType {
 }
 
 export const maxCommentsForFocused = 2;
-	 
+
+export const findLastMonday = (): string => {
+	const today: Date = new Date();
+	const dayOfWeek: number = today.getDay(); // 0 for Sunday, 1 for Monday, and so on
+	const diff: number =
+		today.getDate() - dayOfWeek + (dayOfWeek === 0 ? -6 : 1); // Adjust to get the last Monday
+	const lastMonday: Date = new Date(today.setDate(diff));
+	const year = lastMonday.getFullYear().toString().slice(-2); // Extract last two digits of the year
+	let month: string | number = lastMonday.getMonth() + 1;
+	month = month < 10 ? '0' + month : month;
+	let day: string | number = lastMonday.getDate();
+	day = day < 10 ? '0' + day : day;
+	return `${day}-${month}-${year}`;
+};
